ignore existing gotty

+++ shell/latest.sh
@@ -9,9 +9,11 @@ export GOTTY_SESSION="shell"
 #
 # Gotty
 #
-wget -O /usr/local/bin/gotty https://raw.githubusercontent.com/portalplatform/apps/master/gotty/gotty
-chmod 755 /usr/local/bin/gotty
-setcap cap_net_bind_service=+ep /usr/local/bin/gotty
+if [ ! -e /usr/local/bin/gotty ] ; then
+    wget -O /usr/local/bin/gotty https://raw.githubusercontent.com/portalplatform/apps/master/gotty/gotty
+    chmod 755 /usr/local/bin/gotty
+    setcap cap_net_bind_service=+ep /usr/local/bin/gotty
+fi
 
 
 #
